replaced BaseBuffer with ChannelManager

+++ src/liblvr2/display/PointCloud.cpp
@@ -84,7 +84,7 @@ void PointCloud::init(PointBufferPtr buffer)
         if (m_normals)
             m_numNormals = n_points;
 
-        unsigned w_color, dummy;
+        size_t w_color, dummy;
 		ucharArr colors      = buffer->getColorArray(w_color);
 		floatArr intensities = buffer->getFloatArray("intensities", n_points, dummy);
